[p11] Split PublicKey and PrivateKey into separate files.
